backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
aiosqlite==0.20.*
apscheduler==3.10.*
httpx==0.28.*
pydantic==2.10.*
python-dotenv==1.0.*
